added checkpoints

+++ ShippingApp/Models/AddProductTypeRequest.cs
@@ -3,7 +3,7 @@
     public class AddProductTypeRequest
     {
         public string type { get; set; }
-        public decimal price { get; set; }
+        public float price { get; set; }
         public bool isFragile { get; set; }
     }
 }
