+++ ShippingApp/Models/ProductType.cs
@@ -4,7 +4,7 @@
     {
         public Guid productTypeId { get; set; }
         public string type { get; set; }
-        public decimal price { get; set; } = -1;
+        public float price { get; set; } = -1;
         public bool isFragile { get; set; } 
     }
 }
